app/requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
nltk==3.8.1
sentence-transformers==3.0.1
weaviate-client==4.5.4
langgraph==0.1.9
langchain==0.2.11
langchain-openai==0.1.7
boto3==1.34.158
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-jaeger==1.21.0
python-dotenv==1.0.1
pytest==8.2.1
mypy==1.10.0
black==24.4.2
slowapi==0.1.9
locust==2.20.1
# Advanced GenAI dependencies
torch==2.0.1
transformers==4.36.2
peft==0.7.1
trl==0.7.4
datasets==2.15.0
llama-cpp-python==0.2.23
dspy==2.4.0
langfuse==2.7.3
